src/main/webapp/app/entities/venda-veiculo/delete/venda-veiculo-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, fakeAsync, inject, tick } from '@angular/core/testing';
import { HttpResponse, provideHttpClient } from '@angular/common/http';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { VendaVeiculoService } from '../service/venda-veiculo.service';

import { VendaVeiculoDeleteDialogComponent } from './venda-veiculo-delete-dialog.component';

describe('VendaVeiculo Management Delete Component', () => {
  let comp: VendaVeiculoDeleteDialogComponent;
  let fixture: ComponentFixture<VendaVeiculoDeleteDialogComponent>;
  let service: VendaVeiculoService;
  let mockActiveModal: NgbActiveModal;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [VendaVeiculoDeleteDialogComponent],
      providers: [provideHttpClient(), NgbActiveModal],
    })
      .overrideTemplate(VendaVeiculoDeleteDialogComponent, '')
      .compileComponents();
    fixture = TestBed.createComponent(VendaVeiculoDeleteDialogComponent);
    comp = fixture.componentInstance;
    service = TestBed.inject(VendaVeiculoService);
    mockActiveModal = TestBed.inject(NgbActiveModal);
  });

  describe('confirmDelete', () => {
    it('Should call delete service on confirmDelete', inject(
      [],
      fakeAsync(() => {
        // GIVEN
        jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({ body: {} })));

        // WHEN
        comp.confirmDelete('ABC');
        tick();

        // THEN
        expect(service.delete).toHaveBeenCalledWith('ABC');
        expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
      }),
    ));

    it('Should not call delete service on clear', () => {
      // GIVEN
      jest.spyOn(service, 'delete');

      // WHEN
      comp.cancel();

      // THEN
      expect(service.delete).not.toHaveBeenCalled();
      expect(mockActiveModal.close).not.toHaveBeenCalled();
      expect(mockActiveModal.dismiss).toHaveBeenCalled();
    });
  });
});
